Added extractPharmacyDispenseMessage stubs for keys
