+++ api/src/main/java/org/openmrs/module/kenyaemrIL/il/pharmacy/PHARMACY_DISPENSE.java
@@ -1,18 +1,27 @@
 package org.openmrs.module.kenyaemrIL.il.pharmacy;
+
 /**
  * @author Stanslaus Odhiambo
  *         Created on 21/11/2017.
  */
-public class PharmacyEncodedOrder {
+public class PHARMACY_DISPENSE {
+    private String dispenseNotes;
     private String frequency;
-    private String prescriptionNotes;
+    private String quantityDispensed;
     private String dosage;
     private String codingSystem;
-    private String quantityPrescribed;
     private String strength;
     private String duration;
+    private String actualDrugs;
     private String drugName;
 
+    public String getDispenseNotes() {
+        return dispenseNotes;
+    }
+
+    public void setDispenseNotes(String dispenseNotes) {
+        this.dispenseNotes = dispenseNotes;
+    }
 
     public String getFrequency() {
         return frequency;
@@ -22,12 +31,12 @@ public class PharmacyEncodedOrder {
         this.frequency = frequency;
     }
 
-    public String getPrescriptionNotes() {
-        return prescriptionNotes;
+    public String getQuantityDispensed() {
+        return quantityDispensed;
     }
 
-    public void setPrescriptionNotes(String prescriptionNotes) {
-        this.prescriptionNotes = prescriptionNotes;
+    public void setQuantityDispensed(String quantityDispensed) {
+        this.quantityDispensed = quantityDispensed;
     }
 
     public String getDosage() {
@@ -46,14 +55,6 @@ public class PharmacyEncodedOrder {
         this.codingSystem = codingSystem;
     }
 
-    public String getQuantityPrescribed() {
-        return quantityPrescribed;
-    }
-
-    public void setQuantityPrescribed(String quantityPrescribed) {
-        this.quantityPrescribed = quantityPrescribed;
-    }
-
     public String getStrength() {
         return strength;
     }
@@ -68,6 +69,14 @@ public class PharmacyEncodedOrder {
 
     public void setDuration(String duration) {
         this.duration = duration;
+    }
+
+    public String getActualDrugs() {
+        return actualDrugs;
+    }
+
+    public void setActualDrugs(String actualDrugs) {
+        this.actualDrugs = actualDrugs;
     }
 
     public String getDrugName() {
