+++ api/src/main/java/org/openmrs/module/kenyaemrIL/il/pharmacy/PHARMACY_ENCODED_ORDER.java
@@ -1,27 +1,18 @@
 package org.openmrs.module.kenyaemrIL.il.pharmacy;
-
 /**
  * @author Stanslaus Odhiambo
  *         Created on 21/11/2017.
  */
-public class PharmacyDispense {
-    private String dispenseNotes;
+public class PHARMACY_ENCODED_ORDER {
     private String frequency;
-    private String quantityDispensed;
+    private String prescriptionNotes;
     private String dosage;
     private String codingSystem;
+    private String quantityPrescribed;
     private String strength;
     private String duration;
-    private String actualDrugs;
     private String drugName;
 
-    public String getDispenseNotes() {
-        return dispenseNotes;
-    }
-
-    public void setDispenseNotes(String dispenseNotes) {
-        this.dispenseNotes = dispenseNotes;
-    }
 
     public String getFrequency() {
         return frequency;
@@ -31,12 +22,12 @@ public class PharmacyDispense {
         this.frequency = frequency;
     }
 
-    public String getQuantityDispensed() {
-        return quantityDispensed;
+    public String getPrescriptionNotes() {
+        return prescriptionNotes;
     }
 
-    public void setQuantityDispensed(String quantityDispensed) {
-        this.quantityDispensed = quantityDispensed;
+    public void setPrescriptionNotes(String prescriptionNotes) {
+        this.prescriptionNotes = prescriptionNotes;
     }
 
     public String getDosage() {
@@ -55,6 +46,14 @@ public class PharmacyDispense {
         this.codingSystem = codingSystem;
     }
 
+    public String getQuantityPrescribed() {
+        return quantityPrescribed;
+    }
+
+    public void setQuantityPrescribed(String quantityPrescribed) {
+        this.quantityPrescribed = quantityPrescribed;
+    }
+
     public String getStrength() {
         return strength;
     }
@@ -69,14 +68,6 @@ public class PharmacyDispense {
 
     public void setDuration(String duration) {
         this.duration = duration;
-    }
-
-    public String getActualDrugs() {
-        return actualDrugs;
-    }
-
-    public void setActualDrugs(String actualDrugs) {
-        this.actualDrugs = actualDrugs;
     }
 
     public String getDrugName() {
